requirements.txt
google-genai
streamlit
numpy
scikit-learn
pandas
plotly
ollama